fix(contracts): epoch length, artifacts path
