+++ packages/contracts/src/deploy.ts
@@ -1,7 +1,5 @@
-import { BigNumberish, ethers, utils } from 'ethers'
+import { BigNumberish, ethers } from 'ethers'
 import path from 'path'
-import fs from 'fs'
-import solc from 'solc'
 import {
     Circuit,
     EPOCH_TREE_DEPTH,
@@ -15,121 +13,16 @@ import {
 import { Unirep, Unirep__factory as UnirepFactory } from '../typechain'
 import poseidon from '../src/poseidon'
 import { ATTESTING_FEE, EPOCH_LENGTH } from './config'
+import { compileVerifier, createVerifierName, linkLibrary } from './utils'
 
-export { Unirep, UnirepFactory }
-
-function linkLibrary(
-    bytecode: string,
-    libraries: {
-        [name: string]: string
-    } = {}
-): string {
-    let linkedBytecode = bytecode
-    for (const [name, address] of Object.entries(libraries)) {
-        const placeholder = `__\$${utils
-            .solidityKeccak256(['string'], [name])
-            .slice(2, 36)}\$__`
-        const formattedAddress = utils
-            .getAddress(address)
-            .toLowerCase()
-            .replace('0x', '')
-        if (linkedBytecode.indexOf(placeholder) === -1) {
-            throw new Error(`Unable to find placeholder for library ${name}`)
-        }
-        while (linkedBytecode.indexOf(placeholder) !== -1) {
-            linkedBytecode = linkedBytecode.replace(
-                placeholder,
-                formattedAddress
-            )
-        }
+function tryPath(file: string) {
+    let artifacts: any
+    try {
+        artifacts = require(path.join(__dirname, '../build/artifacts', file))
+    } catch (_) {
+        artifacts = require(path.join(__dirname, '../artifacts', file))
     }
-    return linkedBytecode
-}
-
-export const createVerifierName = (circuit: Circuit | string) => {
-    return `${circuit.charAt(0).toUpperCase() + circuit.slice(1)}Verifier`
-}
-
-export const genVerifier = (contractName: string, vk: any): string => {
-    const templatePath = path.resolve(
-        __dirname,
-        './template/groth16Verifier.txt'
-    )
-
-    let template = fs.readFileSync(templatePath, 'utf8')
-
-    template = template.replace('<%contract_name%>', contractName)
-
-    const vkalpha1 =
-        `uint256(${vk.vk_alpha_1[0].toString()}),` +
-        `uint256(${vk.vk_alpha_1[1].toString()})`
-    template = template.replace('<%vk_alpha1%>', vkalpha1)
-
-    const vkbeta2 =
-        `[uint256(${vk.vk_beta_2[0][1].toString()}),` +
-        `uint256(${vk.vk_beta_2[0][0].toString()})], ` +
-        `[uint256(${vk.vk_beta_2[1][1].toString()}),` +
-        `uint256(${vk.vk_beta_2[1][0].toString()})]`
-    template = template.replace('<%vk_beta2%>', vkbeta2)
-
-    const vkgamma2 =
-        `[uint256(${vk.vk_gamma_2[0][1].toString()}),` +
-        `uint256(${vk.vk_gamma_2[0][0].toString()})], ` +
-        `[uint256(${vk.vk_gamma_2[1][1].toString()}),` +
-        `uint256(${vk.vk_gamma_2[1][0].toString()})]`
-    template = template.replace('<%vk_gamma2%>', vkgamma2)
-
-    const vkdelta2 =
-        `[uint256(${vk.vk_delta_2[0][1].toString()}),` +
-        `uint256(${vk.vk_delta_2[0][0].toString()})], ` +
-        `[uint256(${vk.vk_delta_2[1][1].toString()}),` +
-        `uint256(${vk.vk_delta_2[1][0].toString()})]`
-    template = template.replace('<%vk_delta2%>', vkdelta2)
-
-    template = template.replace(
-        '<%vk_input_length%>',
-        (vk.IC.length - 1).toString()
-    )
-    template = template.replace('<%vk_ic_length%>', vk.IC.length.toString())
-    let vi = ''
-    for (let i = 0; i < vk.IC.length; i++) {
-        if (vi.length !== 0) {
-            vi = vi + '        '
-        }
-        vi =
-            vi +
-            `vk.IC[${i}] = Pairing.G1Point(uint256(${vk.IC[
-                i
-            ][0].toString()}),` +
-            `uint256(${vk.IC[i][1].toString()}));\n`
-    }
-    template = template.replace('<%vk_ic_pts%>', vi)
-
-    return template
-}
-
-export const compileVerifier = async (contractName: string, vkey: any) => {
-    const fileName = contractName + '.sol'
-    const sources = {}
-    sources[fileName] = {}
-    sources[fileName]['content'] = genVerifier(contractName, vkey)
-    const input = {
-        language: 'Solidity',
-        sources: sources,
-        settings: {
-            outputSelection: {
-                '*': {
-                    '*': ['*'],
-                },
-            },
-        },
-    }
-
-    const output = JSON.parse(solc.compile(JSON.stringify(input)))
-    return {
-        abi: output.contracts[fileName][contractName].abi,
-        bytecode: output.contracts[fileName][contractName].evm.bytecode.object,
-    }
+    return artifacts
 }
 
 /**
@@ -180,30 +73,11 @@ export const deployUnirep = async (
         await c.deployed()
         libraries[`Poseidon${inputCount}`] = c.address
     }
-    let incArtifacts: any
-    try {
-        incArtifacts = require(path.join(
-            __dirname,
-            '../build/artifacts/@zk-kit/incremental-merkle-tree.sol/IncrementalBinaryTree.sol/IncrementalBinaryTree.json'
-        ))
-    } catch (_) {
-        incArtifacts = require(path.join(
-            __dirname,
-            '../artifacts/@zk-kit/incremental-merkle-tree.sol/IncrementalBinaryTree.sol/IncrementalBinaryTree.json'
-        ))
-    }
-    let smtArtifacts: any
-    try {
-        smtArtifacts = require(path.join(
-            __dirname,
-            '../build/artifacts/contracts/SparseMerkleTree.sol/SparseMerkleTree.json'
-        ))
-    } catch (_) {
-        smtArtifacts = require(path.join(
-            __dirname,
-            '../artifacts/contracts/SparseMerkleTree.sol/SparseMerkleTree.json'
-        ))
-    }
+    const incPath =
+        '@zk-kit/incremental-merkle-tree.sol/IncrementalBinaryTree.sol/IncrementalBinaryTree.json'
+    const incArtifacts: any = tryPath(incPath)
+    const smtPath = 'contracts/SparseMerkleTree.sol/SparseMerkleTree.json'
+    const smtArtifacts: any = tryPath(smtPath)
     const incrementalMerkleTreeFactory = new ethers.ContractFactory(
         incArtifacts.abi,
         linkLibrary(incArtifacts.bytecode, {
@@ -231,10 +105,8 @@ export const deployUnirep = async (
         console.log(`Deploying ${contractName}`)
         let artifacts
         if (prover === undefined) {
-            artifacts = require(path.join(
-                __dirname,
-                `../build/artifacts/contracts/verifiers/${contractName}.sol/${contractName}.json`
-            ))
+            const verifierPath = `contracts/verifiers/${contractName}.sol/${contractName}.json`
+            artifacts = tryPath(verifierPath)
         } else {
             const vkey = prover.getVKey(circuit)
             artifacts = await compileVerifier(contractName, vkey)
