+++ packages/core/test/synchronizer/synchronizer.test.ts
@@ -1,25 +1,14 @@
 import { ethers } from 'hardhat'
-import { BigNumber } from 'ethers'
 import { expect } from 'chai'
 import { ZkIdentity, hashLeftRight } from '@unirep/crypto'
-import { EPOCH_LENGTH } from '@unirep/circuits'
+import { EPOCH_LENGTH } from '@unirep/contracts'
 import { defaultProver } from '@unirep/circuits/provers/defaultProver'
 import { deployUnirep } from '@unirep/contracts/deploy'
 
 const attestingFee = ethers.utils.parseEther('0.1')
 
-import {
-    Synchronizer,
-    schema,
-    decodeBigIntArray,
-    computeInitUserStateRoot,
-} from '../../src'
-import {
-    genUserState,
-    genRandomAttestation,
-    compareDB,
-    submitUSTProofs,
-} from '../utils'
+import { Synchronizer, schema, computeInitUserStateRoot } from '../../src'
+import { genUserState, compareDB, submitUSTProofs } from '../utils'
 import { SQLiteConnector } from 'anondb/node'
 
 let synchronizer: Synchronizer
